requirements.txt
PyOpenGL
glfw
numpy
PyGLM
pillow